Move drone_compatibility to module pipeline, set drone envirinment vars before executing steps
