+++ pipeline/frontend/metadata.go
@@ -246,8 +246,6 @@ func (m *Metadata) Environ() map[string]string {
 		params["CI_COMMIT_PULL_REQUEST"] = pullRegexp.FindString(m.Curr.Commit.Ref)
 	}
 
-	m.setDroneEnviron(params)
-
 	return params
 }
